gas/riscv: Don't error when getting the isa for the 'e' extension
Previously a fatal error would be emitted if the user simply tried
to query for the corresponding ISA for the 'e' base when the xlen
was >32. This check has now been moved so that it is only an error
when the user tries to *use* the corresponding ISA for the 'e' base.

gas/ChangeLog:

	* config/tc-riscv.c (riscv_isa_for_subset): Do not
	emit an error when retrieving the ISA for the 'e' extension
	when xlen is >32. This is checked elsewhere.