opcodes/riscv: Don't disassemble rv32 instructions if ELF mach is rv64
+ vica versa

Previously objdump would just disassemble the superset of both the
riscv32 and riscv64 instruction sets, which didn't work when there
were decode conflicts for instructions.

Now, based on the bfd mach we either disassemble for the riscv32 or
riscv64 instruction set (+ all of the extensions).

Note: This change involves hardcoded the extensions supported for
each machine in cpu/riscv.opc, however this information is already
available in the (define-mach) statements in cpu/riscv.cpu. It would
be good to remove this redundancy.

cpu/ChangeLog:

	* riscv.opc (disassemble_init_riscv): Added function to set
	up the supported ISAs based on the machine id.
	(CGEN_VALIDATE_INSN_SUPPORTED): Define macro so that
	riscv_cgen_insn_supported will be called to check before
	disassembling to a specific instruction.

include/ChangeLog:

	* include/dis-asm.h (disassemble_init_riscv): Added forward
	declaration.

opcodes/ChangeLog:

	* disassemble.c (disassemble_init_for_target): Call
	disassemble_init_riscv to set correct ISA at init time.
	* riscv-dis.c: Regenerated.
	* riscv-opc.h: Likewise.