opcodes/riscv: Add support for .insn in cpu description
The special .insn directive is handled by adding a number of
special instructions which are not intended to be used directly
by a user. In the cgen cpu description, these have the form

  _insn_<mnemonic> <opcode>, ... operands ...

Which will correspond to the assembler directive

  .insn <mnemonic> <opcode>, ... operands ...

The assembler will morph the .insn directive into a form which
the cgen generated code can handle.

cpu/ChangeLog:

	* riscv.cpu: Add new definitions for hardware, operands, and
	instructions to supported the special .insn assembler
	directive.
	* riscv.opc (parse_opcode7 parse_copcode2): Add handlers to
	parse enumeration values of opcodes in instructions specified
	through the .insn assembler directive.

opcodes/ChangeLog:

	* riscv-asm.c: Regenerated.
	* riscv-desc.c: Likewise.
	* riscv-desc.h: Likewise.
	* riscv-dis.c: Likewise.
	* riscv-ibld.c: Likewise.
	* riscv-opc.c: Likewise.
	* riscv-opc.h: Likewise.