Add compressed instructions into simulator support
This reintroduces the first of the compressed instructions
into the simulator and regenerated opcodes and the simulator.

NB: Generating the decoders for the simulators requires
a patch to cgen, which means top-of-tree will no longer
work. At the time of this commit these changes were
in erj-fixes, however a patch is also included in this commit
(cgen-sim-decoder.patch).