cpu/riscv: Add 'A' and 'M' instruction semantics
The semantics for the A extension instructions are not correct
at the moment, but should work in the simplest of execution
environments (single core, single process).

The mulh, mulhsu mulhu instructions have semantics whose
implementations differ considerably between the 32 and 64-bit
implementations. This is because when xlen==64 these instructions
need to read the upper half of a 128-bit output but there
is no way to express a result that large in the cgen rtl. As a
consequence these 64x64 multiplications must be composed of
many narrower operations, which results in a very complicated
implementation for riscv64.

The easiest way to allow both a simple riscv32 implementation
and complication riscv64 implementation was to implement both
and then switch behavior at runtime. This required a way to
determine xlen at runtime by reading a special virtual register
h-xlen from the rtl.

cpu/ChangeLog:

	* riscv.cpu: Add semantics for all 'M' and 'A' instructions.
	(h-xlen): Define a hardware register which represents the
	xlen of the machine. This is used by some 'M' instructions
	where the semantic implementation differs considerably between
	the 32-bit and 64-bit machines.