sim/riscv: Implement brk syscall emulation
This adds custom emulation of the brk syscall. This works by
allocating a new memory region to hold a fixed size heap (16MiB),
and then the simulator keeps track of the current program break
in this heap.

In addition, the stack has been moved so that it descends from
0x7fffffff instead of 0xffffffff. The heap then meets at the
end of the stack, and starts from 16MiB earlier. So the default
memory layout now looks as follows:

0x7fffffff-0x7ff00000   stack
0x7fefffff-0x7fb00000   heap
...
0x01f00000-0x00000000   program

The stack pointer is now initialized to 0x7ffffff0