riscv: Unify include/gdb/sim-riscv.h and gdb/riscv-tdep.h
To avoid having two copies of the regsiter table in gdb/riscv-tdep.h in
include/gdb/sim-riscv.h this commit:
  * copies the latest version of the table from riscv-tdep.h to
    sim-riscv.h.
  * deletes the table from riscv-tdep.h and includes sim-riscv.h instead

At the moment, the register table is all that sim-riscv.h contains. If in
future the header is extended, we could consider moving the register table
to a separate header file, so that extraneous code from sim-riscv.h does
not get included in riscv-tdep.h.