sim: Fix underflow detection for non-"near even" rounding
This attempts to fix the detection of underflow for rounding modes
other than "near even". There is an edge case where values with
an unbounded exponent would round to a subnormal number, but after
rounding they end up being normal (non-subnormal) numbers. This
is a problem because these numbers *should* be detected as *tiny*, but
if tininess is only detected after rounding, after the exponent
been normalized, this edge case is missed.

I'm still not convinced that these fixes are perfect, but they
do improve results when validated with TestFloat.

sim/common/ChangeLog:

	* sim-fpu.c (do_normal_round): Correct the setting of
	the underflow flag (detection of tininess) for non-"near even"
	rounding modes.