sim: Fix reporting of exceptions from comparison operations
This fixes ltsf, lesf, gtsf, gesf, ltdf, ledf, gtdf and gedf so
that they now signal any exceptions using the FPU error handler.
This can happen when an input to these comparison options
is a NaN.

The operations have been changed so that the above comparisons
default to signalling comparisons, but eqsf, nesf, eqdf and nedf
are still non-signalling. This was chosen based on the behaviour
of RISC-V instructions, however in future this should be fixed so
that it is configurable whether or not a given comparison
operation is signalling.

sim/common/ChangeLog:

	* cgen-accfp.c (ltsf lesf gtsf gesf ltdf ledf gtdf gedf)
	Update functions to use the FPU error handler to report
	exceptions from the floating point comparison operations.
	Use the entry points to the FPU which return exceptions
	rather than those which suppress them.