[PATCH 1/2] RISC-V: Add cgen cpu description and generated files
This adds a cpu definition for RISC-V as well as the
cgen-generated opcodes files. This does not replace
the existing opcodes files and it has no effect on the
RISC-V binutils at this time except to include cgen
dependencies and generated files in the libopcodes build.
These files lay the foundation for a RISC-V cgen-based
simulator.

cpu/ChangeLog:

	* riscv.cpu: New file.
	* riscv.opc: New file.

opcodes/ChangeLog:

	* Makefile.am: Add rules for the files generated from
	cpu/riscv.cpu and cpu/riscv.opc.
	* Makefile.in: Regenerated.
	* configure: Regenerated.
	* configure.ac (bfd_riscv_arch): Specify that cgen is used and
	add cgen generated dependencies.
	* riscv-cgen-desc.c: Generated.
	* riscv-cgen-desc.h: Likewise.
	* riscv-cgen-opc.c: Likewise.
	* riscv-cgen-opc.h: Likewise.