Fix bug in riscv_cgen_insn_supported
This checks the CGEN_INSN_RVEXT attribute to see whether
a given instruction is supported however it was using
an 'int' to hold the mask. There are over 32 extensions
now so an 'int' is not guaranteed to be enough to hold a
mask for all of them.

cpu/ChangeLog:

	* riscv.opc (riscv_cgen_insn_supported): Use long long
	for type of extension attribute values as there are
	too many extensions for the mask to fix in a 32-bit
	value now.

opcodes/ChangeLog:

	* riscv-cgen-opc.c: Regenerated.