riscv: Attempt to add rve support
This adds the --riscv-rve flag to the simulator which swaps out the
register used for syscall identification, such that rve binaries
can correctly exit.

If possible this should be swapped out for (for example), checking the
e_flags field of an ELF file/integrating into the sim model when
merging with newer cgen codebases.